Moved warmup kernel to seperate file ( initMC -> cudaFunction.cc ). Moved cudaFunctions to a c file from a header.
